+++ tests/apache/hive/hooks/test_hive.py
@@ -56,3 +56,44 @@ async def test_partition_exists(mock_get_client, mock_get_connection, result, re
     cursor.fetchall.return_value = result
     res = await hook.partition_exists(TEST_TABLE, TEST_SCHEMA, TEST_PARTITION, TEST_POLLING_INTERVAL)
     assert res == response
+
+
+@pytest.mark.parametrize(
+    "partition,expected",
+    [
+        ("user_profile/city=delhi", ("default", "user_profile", "city=delhi")),
+        ("user.user_profile/city=delhi", ("user", "user_profile", "city=delhi")),
+    ],
+)
+def test_parse_partition_name_success(partition, expected):
+    """Assert that `parse_partition_name` correctly parse partition string"""
+    actual = HiveCliHookAsync.parse_partition_name(partition)
+    assert actual == expected
+
+
+def test_parse_partition_name_exception():
+    """Assert that `parse_partition_name` throw exception if partition string not correct"""
+    with pytest.raises(ValueError):
+        HiveCliHookAsync.parse_partition_name("user_profile.city=delhi")
+
+
+@pytest.mark.parametrize(
+    "result,expected",
+    [
+        (["123"], True),
+        ([], False),
+    ],
+)
+@mock.patch("astronomer.providers.apache.hive.hooks.hive.HiveCliHookAsync.get_connection")
+@mock.patch("astronomer.providers.apache.hive.hooks.hive.HiveCliHookAsync.get_hive_client")
+def test_check_partition_exists(mock_get_client, mock_get_connection, result, expected):
+    """Assert that `check_partition_exists` return True if partition found else return False."""
+    hook = HiveCliHookAsync(metastore_conn_id=TEST_METASTORE_CONN_ID)
+    hiveserver_connection = mock.AsyncMock(HiveServer2Connection)
+    mock_get_client.return_value = hiveserver_connection
+    cursor = mock.AsyncMock(HiveServer2Cursor)
+    hiveserver_connection.cursor.return_value = cursor
+    cursor.is_executing.return_value = False
+    cursor.fetchall.return_value = result
+    actual = hook.check_partition_exists(TEST_SCHEMA, TEST_TABLE, TEST_PARTITION)
+    assert actual == expected